Move all code from esmvalcore/_data_finder.py to esmvalcore/local.py
