+++ esmvalcore/_recipe_checks.py
@@ -9,8 +9,8 @@ from shutil import which
 import isodate
 import yamale
 
-from ._data_finder import _parse_period, get_start_end_year
 from .exceptions import InputFilesNotFound, RecipeError
+from .local import _get_start_end_year, _parse_period
 from .preprocessor import TIME_PREPROCESSORS, PreprocessingTask
 from .preprocessor._ancillary_vars import PREPROCESSOR_ANCILLARIES
 from .preprocessor._multimodel import STATISTIC_MAPPING
@@ -162,7 +162,7 @@ def data_availability(dataset, log=True):
     available_years = set()
 
     for file in input_files:
-        start, end = get_start_end_year(file.name)
+        start, end = _get_start_end_year(file.name)
         available_years.update(range(start, end + 1))
 
     missing_years = required_years - available_years
