+++ tests/unit/local/test_select_files.py
@@ -1,4 +1,4 @@
-from esmvalcore._data_finder import select_files
+from esmvalcore.local import _select_files
 
 
 def test_select_files():
@@ -10,7 +10,7 @@ def test_select_files():
         "pr_Amon_MPI-ESM1-2-HR_historical_r1i1p1f1_gn_197001-197412.nc",
     ]
 
-    result = select_files(files, '1962/1967')
+    result = _select_files(files, '1962/1967')
 
     expected = [
         "pr_Amon_MPI-ESM1-2-HR_historical_r1i1p1f1_gn_196001-196412.nc",
@@ -29,7 +29,7 @@ def test_select_files_monthly_resolution():
         "pr_Amon_EC-Earth3_dcppA-hindcast_s1960-r1i1p1f1_gr_196311-196410.nc",
     ]
 
-    result = select_files(files, '196201/196205')
+    result = _select_files(files, '196201/196205')
 
     expected = [
         "pr_Amon_EC-Earth3_dcppA-hindcast_s1960-r1i1p1f1_gr_196111-196210.nc"
@@ -48,7 +48,7 @@ def test_select_files_daily_resolution():
         filename + "19621101-19631031.nc"
     ]
 
-    result = select_files(files, '19600101/19611215')
+    result = _select_files(files, '19600101/19611215')
 
     expected = [
         filename + "19601101-19611031.nc",
@@ -73,10 +73,10 @@ def test_select_files_sub_daily_resolution():
         filename + "19621101T0300-19631031T2100.nc",
     ]
 
-    result_no_separator = select_files(
+    result_no_separator = _select_files(
         files_no_separator,
         '19600101T0900/19610101T09HH00MM')
-    result_separator = select_files(
+    result_separator = _select_files(
         files_separator,
         '19600101T0900/19610101T0900')
 
@@ -113,8 +113,8 @@ def test_select_files_time_period():
         filename_datetime + "196211010300-196310312100.nc",
     ]
 
-    result_date = select_files(files_date, '196211/P2Y5M')
-    result_datetime = select_files(files_datetime, '19601101T1300/P1Y0M0DT6H')
+    result_date = _select_files(files_date, '196211/P2Y5M')
+    result_datetime = _select_files(files_datetime, '19601101T1300/P1Y0M0DT6H')
 
     expected_date = [
         filename_date + "196211-196310.nc",
@@ -142,9 +142,9 @@ def test_select_files_varying_format():
         filename + "196211010300-196310312100.nc",
     ]
 
-    result_yearly = select_files(files, '1960/1962')
-    result_monthly = select_files(files, '196011/196210')
-    result_daily = select_files(files, '19601101/19601105')
+    result_yearly = _select_files(files, '1960/1962')
+    result_monthly = _select_files(files, '196011/196210')
+    result_daily = _select_files(files, '19601101/19601105')
 
     assert result_yearly == files
     assert result_monthly == files[0:2]
